rename interfaces to models
