+++ dhost/frontend/src/models/Dapp.ts
